requirements-test.txt
pytest
pytest-timeout
pytest-mock
# pinned to support coverage with threading library
pytest-cov==3.0.0
pytest-repeat
pytest-reraise
pytest-custom_exit_code
pytest-asyncio==0.20.3
mock
black==22.3.0
pytest-split==0.8.0
better_profanity==0.7.0
pandas==1.3.5
streamlit==1.16.0
extra-streamlit-components==0.1.55
altair==4.2.2
click==8.0.0
fastapi==0.78.0
uvicorn==0.17.6
uvloop==0.16.0
pydantic==1.9.2
av==9.1.1
requests-mock==1.10.0
pytest-asyncio==0.20.3